requirements.txt
# Core dependencies
numpy==1.24.3
pandas==2.0.2
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
pydantic==2.0.3
python-dotenv==1.0.0

# API and networking
requests-cache==1.0.1
aiohttp==3.8.5
backoff==2.2.1
urllib3==2.0.4

# Data manipulation and analysis
scipy==1.10.1
statsmodels==0.14.0
pandas-ta==0.3.14b0
ta-lib==0.4.26
fastparquet==2023.7.0

# Machine learning and prediction
scikit-learn==1.3.0
tensorflow==2.13.0
keras==2.13.1
xgboost==1.7.6
lightgbm==4.0.0
prophet==1.1.4
joblib==1.3.1

# Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0
dash==2.11.1

# Testing and quality assurance
pytest==7.4.0
pytest-cov==4.1.0
pytest-mock==3.11.1
bandit==1.7.5
flake8==6.0.0
mypy==1.4.1

# Development and tooling
black==23.7.0
isort==5.12.0
pre-commit==3.3.3
jupyter==1.0.0
ipython==8.14.0

requests>=2.31.0
matplotlib>=3.8.0
numpy>=1.24.0
scikit-learn>=1.3.0
